Format code: rename props onHandleChange & rename file InputText
